Auto-format the files with VS and cleanup extra files
I had a duplicate of the manifest file in the src directory, some of the spacing in the TypeScript and XML files was inconsistent, and some of the generated output files were included in the repo.

It might make sense to distribute the generated script files from the repo, but if I'm going to do that it should really be from a dist or public directory to be consistent with other npm packages.