Tidy up some for loops better expressed as array.map
I had a few for loops that iterated over all of the elements in an array, and in some cases it was even transforming elements 1:1 into another array.

A few for loops still make sense in cases where it has an early termination condition. There are also some map calls that output filtered results, so they're still performing an array.push in the callback body.