Get tests working for office-ui-fabric-react components
I needed to add the babel-jest transformer back in, and I need to use react-addons-test-utils instead of the react-test-renderer to get full DOM rendering of those components.

I still need to replace the previous snapshot tests with tests which use the DOM or the component APIs to interogate the state of the UI. The good news is I can now start using simulated events instead of just calling the callbacks from the component props.