Update the column layout and close the dialog immediately
The folder name column is still too wide, adding a maxWidth parameter to that and all of the other columns except Preview to shorten them and let the Preview take as much space as possible. I also made the Preview column multi-line.

As soon as a folder name is clicked and the dialog hands operation back to the UI-less function, we should close the UI and show a progress message indicating that the move is in progress.