Move the excluded folder checks into Data.removeDuplicates
Excluding items in the excluded folder list means we can't find duplicates of those items if the original folder happens to be on the excluded folder list (e.g. the Inbox). Instead of filtering them out of the full list of matches, I'm passing the excluded folderIds in to removeDuplicates and letting it filter down the results after it has already found all of the itemIds in the same folder.