Separate dialog and legacy form UI
Separate ConversationFiler.* into ConversationFilerPage.* and ConversationFilerDialog.*. The ConversationFilerPage component doesn't take stored results from the caller, it needs to load them from the Data.IModel interface, whereas ConversationFilerDialog never loads from the Data.IModel, it must be initialized with the stored results from the UI-less function when it launches the dialog.

To deal with that in the tests, I'm now mocking the Factory module in the ConversationFilerPage tests and returning a mock Data.IModel implementation with test data.

I also prototyped UI-less functions for the aboutDialog and sendFeedback link. I haven't tried those in the manifest yet, but they should replace the out-of-app webpage and mailto: URL I used previously with more integrated implementations.