Fix exception in RESTData when we filter out all folders
The new logic which handles de-duping by leaving the excluded folders in the map is filtering them out from the REST calls to get the folder display names. The .length === 0 check was on the unfiltered list though, so when we try to collate the requests it dereferences the first element and it's undefined.