Add a privacy policy and split bundles for each page
I need a privacy policy statement for the Office Store, so I added one.

The single bundle.js was getting very large, and it had some content that was specific to either index.html or dialog.html (neither of which should be loaded with the other). I split the React.js TSX entrypoint for each of the HTML pages into separate bundles, one-per-page.  That way each page only loads what it actually needs.